edited some rules and added new file 'TestConjunctions'
