+++ automates/text_reading/src/main/resources/application.conf
@@ -60,11 +60,11 @@ CommentEngine {
 
 apps {
 
-projectDir = "/local/path/to/automates"
+projectDir = "/Users/alicekwak/repos/automates/automates/text_reading"
 
 
 ;     wdir = "/Users/bsharp/Downloads/PT-stuf"
-  grfnFile = "/home/alexeeva/Downloads/PETPT_GrFN.json"
+  grfnFile = "/Users/alicekwak/repos/automates/automates/text_reading/input/grfn/petpt_grfn.json"
   inputDirectory = "./input/SIR/text"
 
   inputType = "json"
@@ -95,7 +95,7 @@ projectDir = "/local/path/to/automates"
 
 alignment {
   alignerType = "pairwisew2v"
-  w2vPath = "/local/path/to/vectors.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
+  w2vPath = "/Users/alicekwak/repos/automates/automates/text_reading/vectors.txt" // a word embeddings file (e.g., GloVe) with this header: <len_of_vocab><space><embedding_size>, e.g., "6B 50" (no quotes); use of different sets of embeddings will require adjusting alignment similarity thresholds, e.g., commentTextAlignmentScoreThreshold
   relevantArgs = ["variable", "definition"]
 }
